requirements.txt
psycopg2-binary
numpy
ollama
python-dotenv
rich
